ParserMonads

Documentation

parseeee :: [String] -> IO ()

Using item,char,upper,lower

nothing :: Parser a

Parser with empty value

giveFirst :: Parser Char

Funtion which gives first char from input

useFunction :: (Char -> Bool) -> Parser Char

Returns chars packed in Parser, which fulfill the condition gave as true/false function

char :: Char -> Parser Char

Funcion which is using useFunction in order to check identity

upper :: Parser Char

Function which is using useFunction in order to check if is it a upper letter

lower :: Parser Char

Function which is using useFunction in order to check if is it a lower letter

result :: a -> Parser a

Function which returns value packed in Parser

isVowel :: Parser Char

Using useFunction in order to check if char is a vowel